+++ Kernel/System/SysConfig.pm
@@ -2126,7 +2126,9 @@ sub ConfigurationTranslatedGet {
 
     return %{$Cache} if ref $Cache eq 'HASH';
 
-    my @SettingList = $Self->ConfigurationList();
+    my @SettingList = $Self->ConfigurationList(
+        IncludeInvisible => 1,
+    );
 
     my %Result;
 
@@ -2577,7 +2579,9 @@ sub ConfigurationXML2DB {
         return;
     }
 
-    my @SettingList = $Self->ConfigurationList();
+    my @SettingList = $Self->ConfigurationList(
+        IncludeInvisible => 1,
+    );
 
     my $StorableObject = $Kernel::OM->Get('Kernel::System::Storable');
 
@@ -3160,7 +3164,7 @@ Returns:
 sub ConfigurationList {
     my ( $Self, %Param ) = @_;
 
-    return $Kernel::OM->Get('Kernel::System::SysConfig::DB')->DefaultSettingList();
+    return $Kernel::OM->Get('Kernel::System::SysConfig::DB')->DefaultSettingList(%Param);
 }
 
 =head2 ConfigurationInvalidList()
@@ -4293,7 +4297,9 @@ sub ConfigurationSearch {
 
     my $Search = lc $Param{Search};
 
-    my %Settings = $Self->ConfigurationTranslatedGet();
+    my %Settings = $Self->ConfigurationTranslatedGet(
+        IncludeInvisible => $Param{IncludeInvisible},
+    );
 
     my %Result;
 
@@ -6123,7 +6129,9 @@ sub _DefaultSettingAddBulk {
     }
 
     # Get again all settings.
-    @SettingList = $Self->ConfigurationList();
+    @SettingList = $Self->ConfigurationList(
+        IncludeInvisible => 1,
+    );
 
     $Success = $SysConfigDBObject->DefaultSettingVersionBulkAdd(
         Settings    => \%Settings,