+++ scripts/test/SysConfig/Setting.t
@@ -1302,7 +1302,6 @@ for my $Test (@Tests) {
             ChangeBy     => 1,
         },
     },
-
 );
 
 my @SettingDirtyNames;